refactor: re-export rand from protocol
