+++ core/consensus/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2021"
 [dependencies]
 arc-swap = "1.6"
 
-creep = "0.2"
 futures = { version = "0.3", features = ["async-await"] }
 json = "0.12"
 lazy_static = "1.4"
@@ -35,7 +34,6 @@ protocol = { path = "../../protocol", package = "axon-protocol" }
 bit-vec = "0.6"
 criterion = "0.4"
 num-traits = "0.2"
-rand = "0.7"
 
 [features]
 default = []
