+++ core/executor/Cargo.toml
@@ -27,7 +27,6 @@ molecule = "0.7"
 once_cell = "1.17"
 parking_lot = "0.12"
 protocol = { path = "../../protocol", package = "axon-protocol" }
-rand = { version = "0.8", features = ["small_rng"] }
 ripemd = "0.1"
 rlp = "0.5"
 rlp-derive = "0.1"
@@ -47,9 +46,7 @@ criterion = "0.4"
 ethabi = "18.0"
 ethabi-contract = { git = "https://github.com/rust-ethereum/ethabi.git", rev = "7edf185" }
 ethabi-derive = { git = "https://github.com/rust-ethereum/ethabi.git", rev = "7edf185" }
-getrandom = "0.2"
 hashbrown = "0.13"
-rand7 = { package = "rand", version = "0.7" }
 revm = "2.3"
 rlp = "0.5"
 serde_json = "1.0"
