+++ core/interoperation/Cargo.toml
@@ -17,16 +17,12 @@ log = "0.4"
 
 protocol = { path = "../../protocol", package = "axon-protocol" }
 
-[target.'cfg(not(target_arch = "aarch64"))'.dependencies]
-ckb-vm-aot = "0.22"
-
 [dev-dependencies]
 cardano-serialization-lib = "7.0"
 cardano-message-signing = { git = "https://github.com/ashuralyk/message-signing", branch = "rust" }
 ckb-jsonrpc-types = "0.106"
 ed25519-dalek = "1.0"
 ethers-core = "2.0"
-rand = "0.7"
 serde_json = "1.0"
 
 core-rpc-client = { path = "../../core/rpc-client" }
