+++ core/mempool/Cargo.toml
@@ -26,7 +26,6 @@ protocol = { path = "../../protocol", package = "axon-protocol" }
 [dev-dependencies]
 criterion = { version = "0.4", features = ["async_tokio"] }
 parking_lot = "0.12"
-rand = "0.7"
 
 [[bench]]
 harness = false
