+++ core/network/Cargo.toml
@@ -13,7 +13,6 @@ ipnetwork = "0.20"
 log = "0.4"
 parking_lot = "0.12"
 prost = "0.11"
-rand = "0.7"
 rlp = "0.5"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
