+++ protocol/Cargo.toml
@@ -28,6 +28,7 @@ ophelia = "0.3"
 ophelia-secp256k1 = "0.3"
 overlord = "0.4"
 prost = "0.11"
+rand = "0.7"
 rlp = "0.5"
 rlp-derive = "0.1"
 serde = { version = "1.0", features = ["derive"] }
@@ -35,8 +36,6 @@ tokio = { version = "1.25", features = ["full"] }
 trie = { package = "cita_trie", version = "4.0" }
 
 [dev-dependencies]
-getrandom = "0.2"
 hex = "0.4"
-rand = "0.7"
 serde_json = "1.0"
 toml = "0.7"
